passed partId with url to contact page
