+++ components/PartContactForm.jsx
@@ -1,9 +1,111 @@
 
+'use client'
 
-const PartContactForm = () => {
+import { useRouter, useSearchParams } from "next/navigation"
+import { useEffect } from "react"
+import { useFormState, useFormStatus } from "react-dom"
+import { addMessage } from "@/app/actions/addMessage"
+import { useSession } from "next-auth/react"
+import { toast } from "react-toastify"
+import { FaPaperPlane } from "react-icons/fa"
+
+function ContactForm({ part }) {
+  const router = useRouter()
+  const params = useSearchParams()
+  const { data: session } = useSession()
+
+
+
+  const { state, formAction } = useFormState(addMessage)
+
+  if (state === "submitted") {
+    return (
+      <p className="text-center text-green-500">
+        Your message has been sent successfully
+      </p>
+    )
+  }
+
+  useEffect(() => {
+    if (state === "error") {
+      toast.error("An error occurred. Please try again")
+    }
+    if (state === "success") {
+      toast.success("Your message has been sent successfully")
+
+    }
+  }, [state])
+
   return (
-    <div>PartContactForm</div>
+    session ? (
+      <>
+        <div className="bg-white p-6 rounded-lg shadow-md">
+          <h3 className="text-xl font-bold mb-6">Contact Parts Owner</h3>
+          <form action={formAction}>
+            <input type="hidden" name="part" id="part" defaultValue={params} />
+            <input type="hidden" name="receiver" id="receiver" defaultValue={part.part_owner} />
+            <div className="mb-4">
+              <label className="block text-gray-700 text-sm font-bold mb-2" htmlFor="name">
+                Name:
+              </label>
+              <input
+                className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+                id="name"
+                type="text"
+                placeholder="Enter your name"
+                required
+              />
+            </div>
+            <div className="mb-4">
+              <label className="block text-gray-700 text-sm font-bold mb-2" htmlFor="email">
+                Email:
+              </label>
+              <input
+                className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+                id="email"
+                type="email"
+                placeholder="Enter your email"
+                required
+              />
+            </div>
+            <div className="mb-4">
+              <label className="block text-gray-700 text-sm font-bold mb-2" htmlFor="phone">
+                Phone:
+              </label>
+              <input
+                className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+                id="phone"
+                type="text"
+                placeholder="Enter your phone number"
+              />
+            </div>
+            <div className="mb-4">
+              <label className="block text-gray-700 text-sm font-bold mb-2" htmlFor="message">
+                Message:
+              </label>
+              <textarea
+                className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 h-44 focus:outline-none focus:shadow-outline"
+                id="message"
+                placeholder="Enter your message"
+              ></textarea>
+            </div>
+            <div>
+              <button
+                className="bg-blue-500 hover:bg-blue-600 text-white font-bold py-2 px-4 rounded-full w-full focus:outline-none focus:shadow-outline flex items-center justify-center"
+                type="submit"
+              >
+                <FaPaperPlane className="mr-2" /> Send Message
+              </button>
+            </div>
+          </form>
+        </div>
+      </>
+    ) : (
+      <p className="text-center text-gray-500">
+        Please sign in to contact the owner
+      </p>
+    )
   )
 }
 
-export default PartContactForm
+export default ContactForm
